+++ packages/client-node/src/index.ts
@@ -55,6 +55,12 @@ export {
   type ClickHouseAuth,
   type ClickHouseJWTAuth,
   type ClickHouseCredentialsAuth,
+  type ClickHouseTracer,
+  type ClickHouseSpan,
+  type ClickHouseSpanOptions,
+  type ClickHouseSpanAttributes,
+  type ClickHouseSpanStatus,
+  type ClickHouseSpanName,
 } from "@clickhouse/client-common";
 
 /**
@@ -82,6 +88,9 @@ import {
   isRow as isRow_,
   isException as isException_,
   TupleParam as TupleParam_,
+  ClickHouseSpanNames as ClickHouseSpanNames_,
+  ClickHouseSpanStatusCode as ClickHouseSpanStatusCode_,
+  ClickHouseSpanKind as ClickHouseSpanKind_,
   defaultJSONHandling as defaultJSONHandling_,
 } from "@clickhouse/client-common";
 
@@ -105,4 +114,7 @@ export const isRow = isRow_;
 export const isException = isException_;
 export const TupleParam = TupleParam_;
 export type TupleParam = TupleParam_;
+export const ClickHouseSpanNames = ClickHouseSpanNames_;
+export const ClickHouseSpanStatusCode = ClickHouseSpanStatusCode_;
+export const ClickHouseSpanKind = ClickHouseSpanKind_;
 export const defaultJSONHandling = defaultJSONHandling_;